pyproject.toml
[project]
name = "phone-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "twilio>=9.9.1",
    "uvicorn>=0.40.0",
]
